% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{and}
\alias{and}
\title{Combine predicate functions with \code{&&}}
\usage{
and(...)
}
\arguments{
\item{...}{Predicate functions}
}
\value{
A predicate function
}
\description{
Combine predicate functions with \code{&&}
}
\examples{
and(not_null, not_na)(1)
and(not_null, not_na)(NULL)
}
