% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{and_then}
\alias{and_then}
\alias{bind}
\title{Evaluate a maybe producing function on a maybe value}
\usage{
and_then(.m, .f, ...)

bind(.m, .f, ...)
}
\description{
Evaluate a maybe producing function on a maybe value
}
