% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{and_then2}
\alias{and_then2}
\title{Evaluate a binary maybe returning function on two maybe values}
\usage{
and_then2(.m1, .m2, .f, ...)
}
\arguments{
\item{.m1}{A maybe value}

\item{.m2}{A maybe value}

\item{.f}{A binary maybe returning function to apply to the maybe values}

\item{...}{Named arguments for the function \code{.f}}
}
\value{
A maybe value
}
\description{
Evaluate a binary maybe returning function on two maybe values
}
\examples{
and_then2(just(1), just(2), maybe(`+`))
and_then2(nothing(), just(2), maybe(`/`))
}
