% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{and_then3}
\alias{and_then3}
\title{Evaluate a ternary maybe returning function on three maybe values}
\usage{
and_then3(.m1, .m2, .m3, .f, ...)
}
\arguments{
\item{.m1}{A maybe value}

\item{.m2}{A maybe value}

\item{.m3}{A maybe value}

\item{.f}{A ternary maybe returning function to apply to the maybe values}

\item{...}{Named arguments for the function \code{.f}}
}
\value{
A maybe value
}
\description{
Evaluate a ternary maybe returning function on three maybe values
}
\examples{
safe_sum <- maybe(function(x, y, z) sum(x, y, z))

and_then3(just(1), just(2), just(3), safe_sum)
and_then3(nothing(), just(2), just(3), safe_sum)
}
