% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{filter_justs}
\alias{filter_justs}
\alias{cat_maybes}
\title{Filter and unwrap a list of 'Just' values}
\usage{
filter_justs(.l)

cat_maybes(.l)
}
\arguments{
\item{.l}{List of maybe values}
}
\value{
A list of values
}
\description{
Filter and unwrap a list of 'Just' values
}
\examples{
filter_justs(list(just(1), nothing(), just("a")))
}
