% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{filter_map}
\alias{filter_map}
\title{Map a function over a list and filter only 'Just' values}
\usage{
filter_map(.l, .f)
}
\arguments{
\item{.l}{List of values}

\item{.f}{A maybe returning function to apply to the maybe values}

\item{...}{Named arguments for the function \code{.f}}
}
\value{
A list of values
}
\description{
Map a function over a list and filter only 'Just' values
}
\examples{
filter_map(list(-1, 1, "2"), maybe(sqrt))
}
