% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{flatten_maybe}
\alias{flatten_maybe}
\alias{join}
\title{Flatten a nested maybe value}
\usage{
flatten_maybe(.m)

join(.m)
}
\arguments{
\item{.m}{A maybe value}
}
\value{
A maybe value
}
\description{
Flatten a nested maybe value
}
\examples{
just(just(1)) |> flatten_maybe()
just(nothing()) |> flatten_maybe()
just(1) |> flatten_maybe()
nothing() |> flatten_maybe()
}
