% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{from_just}
\alias{from_just}
\title{Unwrap a 'Just' value or throw an error}
\usage{
from_just(.m)
}
\arguments{
\item{.m}{A maybe value}
}
\value{
The unwrapped 'Just' value
}
\description{
Unwrap a 'Just' value or throw an error
}
\examples{
just(1) \%>\% from_just()
}
