% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_maybe}
\alias{is_maybe}
\title{Check if an object is a maybe value}
\usage{
is_maybe(a)
}
\arguments{
\item{a}{Object to check}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check if an object is a maybe value
}
\examples{
is_maybe(1)
is_maybe(just(1))
is_maybe(nothing())
}
