% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{is_nothing}
\alias{is_nothing}
\title{Check if an object is a 'Nothing' value}
\usage{
is_nothing(a)
}
\arguments{
\item{a}{Object to check}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check if an object is a 'Nothing' value
}
\examples{
is_nothing(1)
is_nothing(just(1))
is_nothing(nothing())
}
