% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{just}
\alias{just}
\title{Create a 'Just' variant of a maybe value}
\usage{
just(a)
}
\arguments{
\item{a}{A value to wrap in a 'Just' container}
}
\value{
A 'Just' variant of a maybe value
}
\description{
Create a 'Just' variant of a maybe value
}
\examples{
just(1)
just("hello")
}
