% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{maybe_equal}
\alias{maybe_equal}
\title{Check if two maybe values are equal}
\usage{
maybe_equal(.m1, .m2)
}
\arguments{
\item{.m1}{A maybe value}

\item{.m2}{A maybe value}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
If both values are 'Nothing' variants or both values are 'Just' variants with
identical contents \code{TRUE} will be returned, otherwise \code{FALSE}.
}
\examples{
maybe_equal(just(1), just(1))
maybe_equal(just(1), just(2))
maybe_equal(nothing(), nothing())
}
