% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{maybe_map2}
\alias{maybe_map2}
\title{Evaluate a binary function on two maybe values}
\usage{
maybe_map2(.m1, .m2, .f, ...)
}
\arguments{
\item{.m1}{A maybe value}

\item{.m2}{A maybe value}

\item{.f}{A binary function to apply to the maybe values}

\item{...}{Named arguments for the function \code{.f}}
}
\value{
A maybe value
}
\description{
Evaluate a binary function on two maybe values
}
\examples{
maybe_map2(just(1), just(2), `+`)
maybe_map2(nothing(), just(2), `/`)
}
