% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{not_empty}
\alias{not_empty}
\title{Check if a vector or data frame is empty}
\usage{
not_empty(a)
}
\arguments{
\item{a}{Object to check}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check if a vector or data frame is empty
}
\examples{
not_empty(integer())
not_empty(list())
not_empty(1:10)
not_empty(data.frame())
not_empty(data.frame(a = 1:10))
}
