% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{not_infinite}
\alias{not_infinite}
\title{Check if an object is infinite}
\usage{
not_infinite(a)
}
\arguments{
\item{a}{Object to check}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check if an object is infinite
}
\examples{
not_infinite(Inf)
not_infinite(1)
}
