% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{not_na}
\alias{not_na}
\title{Check if an object is not NA}
\usage{
not_na(a)
}
\arguments{
\item{a}{Object to check}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check if an object is not NA
}
\examples{
not_na(NA)
not_na(1)
}
