% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{not_null}
\alias{not_null}
\title{Check if an object is NULL}
\usage{
not_null(a)
}
\arguments{
\item{a}{Object to check}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check if an object is NULL
}
\examples{
not_null(NULL)
not_null(1)
}
