% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{not_undefined}
\alias{not_undefined}
\title{Check if an object is undefined}
\usage{
not_undefined(a)
}
\arguments{
\item{a}{}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
In this case 'undefined' values include \code{NULL}, \code{NaN}, all \code{NA} variants,
and infinite values.
}
\examples{
not_undefined(NA)
not_undefined(NULL)
not_undefined(1)
}
