% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{nothing}
\alias{nothing}
\title{Create a 'Nothing' variant of a maybe value}
\usage{
nothing()
}
\value{
A 'Nothing' variant of a maybe value
}
\description{
Create a 'Nothing' variant of a maybe value
}
\examples{
nothing()
}
