% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{perhaps}
\alias{perhaps}
\title{Modify a function to return a default value}
\usage{
perhaps(.f, otherwise, allow_warning = FALSE, result = function() TRUE)
}
\description{
Modify a function to return a default value
}
