% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{with_default}
\alias{with_default}
\alias{from_maybe}
\title{Unwrap a maybe value and return a default for 'Nothing'}
\usage{
with_default(.m, default)

from_maybe(.m, default)
}
\description{
Unwrap a maybe value and return a default for 'Nothing'
}
