% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{or}
\alias{or}
\title{Combine predicate functions to check if any are TRUE}
\usage{
or(...)
}
\arguments{
\item{...}{Predicate functions}
}
\value{
A predicate function
}
\description{
Combine predicate functions to check if any are TRUE
}
\examples{
or(not_null, not_na)(1)
or(not_null, not_na)(NULL)
}
