% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{with_default}
\alias{with_default}
\alias{from_maybe}
\title{Unwrap a maybe value or return a default}
\usage{
with_default(.m, default)

from_maybe(.m, default)
}
\arguments{
\item{.m}{A maybe value}

\item{default}{A default value to return if the maybe value is 'Nothing'}
}
\value{
The unwrapped maybe value or the default value
}
\description{
Unwrap a maybe value or return a default
}
\examples{
just(1) \%>\% with_default(default = 0)
nothing() \%>\% with_default(default = 0)
}
